UnixUser id is a long

+++ foilen-infra-resource-apachephp/src/main/java/com/foilen/infra/resource/apachephp/ApachePhpEventHandler.java
@@ -71,7 +71,7 @@ public class ApachePhpEventHandler extends AbstractCommonMethodUpdateEventHandle
         if (proceed) {
 
             UnixUser unixUser = unixUsers.get(0);
-            Integer unixUserId = unixUser.getId();
+            Long unixUserId = unixUser.getId();
 
             // Create an Application
             Application application = new Application();
